Port theme to Lua using nvim_set_hl()
Now that NeoVim 0.7 is released the nvim_set_hl() API is available.
Using this API we can clean up the color scheme a bit, and also make it
faster to load. On my laptop (an X1 Carbon 7th generation) the Vimscript
theme took around 3 milliseconds to load, while the Lua version takes
about 1 millisecond to load. This isn't really noteworthy in isolation,
but we get this for free by using Lua.